fix signed unsigned warnings

+++ test/shared_test/lib_util_test.cpp
@@ -37,8 +37,8 @@ TEST(sscapiTest, SSC_DATARR_test)
     ssc_var_size(data_arr, &n, nullptr);
     EXPECT_EQ(n, 2);
 //    ssc_var_size(data_arr, &n, nullptr);
-    for (size_t i = 0; i < n; i++){
-        double var = ssc_var_get_number(ssc_var_get_var_array(data_arr, (int)i));
+    for (int i = 0; i < n; i++){
+        double var = ssc_var_get_number(ssc_var_get_var_array(data_arr, i));
         EXPECT_EQ(var, 2 + i);
     }
 
@@ -68,9 +68,9 @@ TEST(sscapiTest, SSC_DATMAT_test)
     EXPECT_EQ(n, 2);
     EXPECT_EQ(m, 2);
 
-    for (size_t i = 0; i < n; i++){
-        for (size_t j = 0; j < m; j++){
-            double var = ssc_var_get_number(ssc_var_get_var_matrix(data_mat, (int)i, (int)j));
+    for (int i = 0; i < n; i++){
+        for (int j = 0; j < m; j++){
+            double var = ssc_var_get_number(ssc_var_get_var_matrix(data_mat, i, j));
             EXPECT_EQ(var, 2 + i * n + j);
         }
     }
